apzkr-pzpi-21-1-boichenko-matvii/Task2-IoT/api-bridge-client/requirements.txt
requests==2.31.0
paho-mqtt==1.6.1
pydantic==2.5.2
